+++ app/src/main/java/com/inky/fitnesscalendar/data/measure/Measure.kt
@@ -1,0 +1,9 @@
+package com.inky.fitnesscalendar.data.measure
+
+import com.inky.fitnesscalendar.ui.util.ContextFormat
+
+interface Measure : ContextFormat {
+    fun isNothing(): Boolean
+}
+
+fun <T : Measure> T.takeIfNotNothing(): T? = takeIf { !isNothing() }