+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+
+# 2025.02.2
+
 - Allow limiting suggested places for an activity type by their color
 - Add dedicated records section to the summary view
 - Don't show statistics if they are 0